// web/src/app/core/api/admin-generated/models/content-batch-list-item-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContentBatchCountsResponse } from '../models/content-batch-counts-response';
import { ContentBatchStatus } from '../models/content-batch-status';
export interface ContentBatchListItemResponse {
  catalogName?: string;
  counts?: ContentBatchCountsResponse;
  createdAt?: string;
  createdByUserId?: string;
  id?: string;
  status?: ContentBatchStatus;
  updatedAt?: string;
}