// web/src/app/core/api/admin-generated/models/content-batch-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContentBatchAuditResponse } from '../models/content-batch-audit-response';
import { ContentBatchCountsResponse } from '../models/content-batch-counts-response';
import { ContentBatchItemResponse } from '../models/content-batch-item-response';
import { ContentBatchStatus } from '../models/content-batch-status';
export interface ContentBatchResponse {
  auditEntries?: Array<ContentBatchAuditResponse>;
  catalogName?: string;
  completedAt?: (string | null);
  counts?: ContentBatchCountsResponse;
  createdAt?: string;
  createdByUserId?: string;
  id?: string;
  items?: Array<ContentBatchItemResponse>;
  startedAt?: (string | null);
  status?: ContentBatchStatus;
  updatedAt?: string;
}